rag-project/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-groq
langchain-chroma
huggingface-hub
langchain-text-splitters
pypdf
chromadb
langchain-huggingface
python-dotenv
python-multipart
tavily-python